requirements.txt
huggingface-hub
sentencepiece
matplotlib
openai
pandas 
seaborn
transformers
sentence-transformers
tqdm
datasets
spacy
load_dotenv
psycopg2-binary
schedule
openpyxl
accelerate
numpy<2
evaluate